Moved logic of HTML generation of scan result to visualization
