+++ python-client/giskard/scanner/result.py
@@ -1,12 +1,6 @@
-from collections import defaultdict
-from html import escape
-from pathlib import Path
-
 import pandas as pd
-from jinja2 import Environment, PackageLoader, select_autoescape
 
 from giskard.utils.analytics_collector import analytics, anonymize
-from ..visualization.custom_jinja import pluralize, format_metric
 
 
 class ScanResult:
@@ -32,46 +26,10 @@ class ScanResult:
         return self.to_html(embed=True)
 
     def to_html(self, filename=None, embed=False):
-        env = Environment(
-            loader=PackageLoader("giskard.visualization", "templates"),
-            autoescape=select_autoescape(),
-        )
-        env.filters["pluralize"] = pluralize
-        env.filters["format_metric"] = format_metric
+        from ..visualization.widget import ScanResultWidget
 
-        tpl = env.get_template("scan_results.html")
-
-        issues_by_group = defaultdict(list)
-        for issue in self.issues:
-            issues_by_group[issue.group].append(issue)
-
-        html = tpl.render(
-            issues=self.issues,
-            issues_by_group=issues_by_group,
-            num_major_issues={
-                group: len([i for i in issues if i.level == "major"]) for group, issues in issues_by_group.items()
-            },
-            num_medium_issues={
-                group: len([i for i in issues if i.level == "medium"]) for group, issues in issues_by_group.items()
-            },
-            num_info_issues={
-                group: len([i for i in issues if i.level == "info"]) for group, issues in issues_by_group.items()
-            },
-        )
-
-        if embed:
-            # Put the HTML in an iframe
-            escaped = escape(html)
-            uid = id(self)
-
-            with Path(__file__).parent.parent.joinpath("templates", "static", "external.js").open("r") as f:
-                js_lib = f.read()
-
-            html = f"""<iframe id="scan-{uid}" srcdoc="{escaped}" style="width: 100%; border: none;" class="gsk-scan"></iframe>
-<script>
-{js_lib}
-(function(){{iFrameResize({{ checkOrigin: false }}, '#scan-{uid}');}})();
-</script>"""
+        widget = ScanResultWidget(self)
+        html = widget.render_html(embed=embed)
 
         if filename is not None:
             with open(filename, "w") as f:
